Implemented custom Debug for InlinableString
I've implemented Debug by calling into the Debug for &str. This means
that it does not show you whether not a string is inline or on the heap.
This could be useful information, but I'm thinking the 99% case is that
the string is being debug printed as part of a bigger structure where
Inline vs Heap is noise, and we only care about the value.