Call deserialize_str instead of deserialize_any
We should hint the Deserializer that we are expecting string, otherwise
it may give us something else and we'll refuse it. For example the
Deserializer from csv crate will attempt to infer type when called with
deserialize_any and if the data contain only digits, assume that it is a
number.